+++ lizard_auth_client/admin.py
@@ -4,3 +4,8 @@ from __future__ import unicode_literals
 from django.contrib import admin
 
 from lizard_auth_client import models
+
+
+admin.site.register(models.Role)
+admin.site.register(models.Organisation)
+admin.site.register(models.UserOrganisationRole)